share user

+++ controller/toDoController.js
@@ -1,65 +1,102 @@
 import toDo from "../models/toDo.js";
-import category from "../models/categories.js";
+import Category from "../models/categories.js";
 
 ///
 
 const createToDo = async (req, res, next) => {
-  const { categoryID, ...other } = req.body;
-  const idCat = req.params.categoryID;
+  const { categoryID } = req.params;
 
   try {
-    const findCategory = await category.find({ _id: idCat });
+    const findCategory = await Category.find({
+      _id: categoryID,
+      $or: [{ user: res.locals.user._id }, { shareUser: res.locals.user._id }],
+    });
+    console.log(findCategory);
 
     if (findCategory.length == 0) {
       res.status(404).send();
       return;
     }
 
-    const NewToDo = await toDo.create({ categoryID: idCat, ...other });
+    const NewToDo = await toDo.create({
+      ...req.body,
+      categoryID: categoryID,
+      createdDate: new Date(),
+    });
     res.json(NewToDo);
   } catch (error) {
     console.log(error);
-    next(error);
+    res.status(424).send({ error: error.message });
   }
 };
 
 ///
 
-const findToDoByID = async (req, res) => {
+const findToDoByID = async (req, res, next) => {
   const { toDoID, categoryID } = req.params;
   try {
-    const toDoById = await toDo.find({ _id: toDoID, categoryID });
-    if (toDoById.length == 0) {
-      res.status(204).send();
-      return;
+    const checkUser = await Category.exists({
+      categoryID,
+      $or: [{ user: res.locals.user._id }, { shareUser: res.locals.user._id }],
+    });
+
+    if (!checkUser) {
+      res.status(404).send();
+    } else {
+      const toDoById = await toDo.find({ _id: toDoID, categoryID });
+      if (toDoById.length == 0) {
+        res.status(204).send();
+        return;
+      }
+      res.send({ toDoById });
     }
-    res.send({ toDoById });
   } catch (error) {
-    res.status(404).send();
+    console.log(error);
+    res.status(424).send({ error: error.message });
   }
 };
 
 ///
 
-const deleteToDoByID = async (req, res) => {
+const deleteToDoByID = async (req, res, next) => {
   const { toDoID, categoryID } = req.params;
   try {
-    await toDo.deleteOne({ _id: toDoID, categoryID });
-  } catch (error) {
-    res.status(404).send();
-    return;
-  }
+    const checkUser = await Category.exists({
+      categoryID,
+      $or: [{ user: res.locals.user._id }, { shareUser: res.locals.user._id }],
+    });
 
-  res.status(204).json({ message: "categoryis removed" });
+    if (!checkUser) {
+      res.status(404).send();
+    } else {
+      const deleteById = await toDo.deleteOne({ _id: toDoID, categoryID });
+      if (deleteById.deletedCount == 0) {
+        res.status(404).send();
+      } else {
+        res.json({ deleteById });
+      }
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(424).send({ error: error.message });
+  }
 };
 
 ///
 
-const updateToDoByID = async (req, res) => {
+const updateToDoByID = async (req, res, next) => {
   const idCat = req.params.categoryID;
   const { toDoID } = req.params;
   const { categoryID, ...other } = req.body;
   try {
+    const checkUser = await Category.exists({
+      categoryID,
+      $or: [{ user: res.locals.user._id }, { shareUser: res.locals.user._id }],
+    });
+
+    if (!checkUser) {
+      res.status(404).send();
+    }
     const updateByID = await toDo.findOneAndUpdate(
       { categoryID: idCat, _id: toDoID },
       { ...other },
@@ -70,22 +107,61 @@ const updateToDoByID = async (req, res) => {
       return;
     }
     res.status(200).json({ updateByID });
-  } catch (error) {}
+  } catch (error) {
+    console.log(error);
+    res.status(424).send({ error: error.message });
+  }
 };
 
 ///
 
-const listToDos = async (req, res) => {
+const listToDos = async (req, res, next) => {
   const { categoryID } = req.params;
-  const { skip, limit } = req.query;
+
+  const { skip, limit, startdate, finishdate, completetask } = req.query;
+  let duoDate;
+  if (startdate && finishdate) {
+    duoDate = { $gte: new Date(startdate), $lte: new Date(finishdate) };
+  } else if (startdate) {
+    duoDate = { $gte: new Date(startdate) };
+  } else if (finishdate) {
+    duoDate = { $lte: new Date(finishdate) };
+  }
+
   try {
-    const ToDoListAll = await toDo.find({ categoryID }).skip(skip).limit(limit);
-    if (ToDoListAll.length == 0) {
-      return res.status(204).send();
+    const checkUser = await Category.exists({
+      categoryID,
+      $or: [{ user: res.locals.user._id }, { shareUser: res.locals.user._id }],
+    });
+    console.log(checkUser);
+    if (!checkUser) {
+      res.status(404).send();
+    } else {
+      let ToDoListAll;
+      if (duoDate) {
+        ToDoListAll = await toDo
+          .find({
+            categoryID,
+            duoDate,
+          })
+          .skip(skip)
+          .limit(limit);
+      } else {
+        ToDoListAll = await toDo
+          .find({
+            categoryID,
+          })
+          .skip(skip)
+          .limit(limit);
+      }
+      if (ToDoListAll.length == 0 || checkUser == false) {
+        return res.status(204).send();
+      }
+      res.status(200).json({ ToDoListAll });
     }
-    res.status(200).json({ ToDoListAll });
   } catch (error) {
-    return res.status(404).send();
+    console.log(error);
+    res.status(424).send({ error: error.message });
   }
 };
